chore: Split an implement for Unified plugin
